Move tests from partests to junit
Parser combinator tests didn't need all the machinery
of partest and it was just slowing them down. This
commit converts all the partest tests into junit.
The sbt project file is modified to get rid of all
the partest cruft.
