+++ src/test/scala/scala/util/parsing/combinator/JsonTest.scala
@@ -1,11 +1,10 @@
-/*
- * filter: inliner warning\(s\); re-run with -Yinline-warnings for details
- */
 import scala.util.parsing.json._
 import scala.collection.immutable.TreeMap
 
-@deprecated("Suppress warnings", since="2.11")
-object Test extends App {
+import org.junit.Test
+import org.junit.Assert.{assertEquals, assertTrue}
+
+class JsonTest {
   /* This method converts parsed JSON back into real JSON notation with objects in
    * sorted-key order. Not required by the spec, but it allows us to do a stable
    * toString comparison. */
@@ -32,12 +31,8 @@ object Test extends App {
   }
 
   // For this one, just parsing should be considered a pass
-  def printJSON(given : String) {
-    JSON parseRaw given match {
-      case None => println("Parse failed for \"%s\"".format(given))
-      case Some(parsed) => println("Passed parse  : " + sortJSON(parsed))
-    }
-  }
+  def printJSON(given : String) : Unit =
+    assertTrue("Parse failed for \"%s\"".format(given), (JSON parseRaw given).isDefined)
    
   // For this usage, do a raw parse (to JSONObject/JSONArray)
   def printJSON(given : String, expected : JSONType) {
@@ -52,18 +47,16 @@ object Test extends App {
   // For this usage, do configurable parsing so that you can do raw if desired
   def printJSON[T](given : String, parser : String => T, expected : Any) {
     parser(given) match {
-      case None => println("Parse failed for \"%s\"".format(given))
-      case Some(parsed) => if (parsed == expected) {
-        println("Passed compare: " + parsed)
-      } else {
+      case None => assertTrue("Parse failed for \"%s\"".format(given), false)
+      case Some(parsed) => if (parsed != expected) {
         val eStr = sortJSON(expected).toString
         val pStr = sortJSON(parsed).toString
-        stringDiff(eStr,pStr)
+        assertTrue(stringDiff(eStr,pStr), false)
       }
     }
   }
 
-  def stringDiff (expected : String, actual : String) {
+  def stringDiff (expected : String, actual : String): String = {
     if (expected != actual) {
       // Figure out where the Strings differ and generate a marker
         val mismatchPosition = expected.toList.zip(actual.toList).indexWhere({case (x,y) => x != y}) match {
@@ -71,57 +64,77 @@ object Test extends App {
           case x => x
         }
         val reason = (" " * mismatchPosition) + "^"
-        println("Expected: %s\nGot     : %s \n          %s".format(expected, actual, reason))
+        "Expected: %s\nGot     : %s \n          %s".format(expected, actual, reason)
 
     } else {
-      println("Passed compare: " + actual)
+      "Passed compare: " + actual
     }
   }
 
-
   // The library should differentiate between lower case "l" and number "1" (ticket #136)
-  printJSON("{\"name\" : \"value\"}", JSONObject(Map("name" -> "value")))
-  printJSON("{\"name\" : \"va1ue\"}", JSONObject(Map("name" -> "va1ue")))
-  printJSON("{\"name\" : { \"name1\" : \"va1ue1\", \"name2\" : \"va1ue2\" } }",
-            JSONObject(Map("name" -> JSONObject(Map("name1" -> "va1ue1", "name2" -> "va1ue2")))))
+  @Test
+  def testEllVsOne: Unit = {
+    printJSON("{\"name\" : \"value\"}", JSONObject(Map("name" -> "value")))
+    printJSON("{\"name\" : \"va1ue\"}", JSONObject(Map("name" -> "va1ue")))
+    printJSON("{\"name\" : { \"name1\" : \"va1ue1\", \"name2\" : \"va1ue2\" } }",
+              JSONObject(Map("name" -> JSONObject(Map("name1" -> "va1ue1", "name2" -> "va1ue2")))))
+  }
 
   // Unicode escapes should be handled properly
-  printJSON("{\"name\" : \"\\u0022\"}")
+  @Test
+  def testEscapes: Unit =
+    printJSON("{\"name\" : \"\\u0022\"}")
 
   // The library should return a map for JSON objects (ticket #873)
-  printJSONFull("{\"function\" : \"add_symbol\"}", Map("function" -> "add_symbol"))
+  @Test
+  def testMap: Unit =
+    printJSONFull("{\"function\" : \"add_symbol\"}", Map("function" -> "add_symbol"))
 
   // The library should recurse into arrays to find objects (ticket #2207)
-  printJSON("[{\"a\" : \"team\"},{\"b\" : 52}]", JSONArray(List(JSONObject(Map("a" -> "team")), JSONObject(Map("b" -> 52.0)))))
+  @Test
+  def testObjectsInArrays: Unit =
+    printJSON("[{\"a\" : \"team\"},{\"b\" : 52}]", JSONArray(List(JSONObject(Map("a" -> "team")), JSONObject(Map("b" -> 52.0)))))
+
   
   // The library should differentiate between empty maps and lists (ticket #3284)
-  printJSONFull("{}", Map()) 
-  printJSONFull("[]", List())
+  @Test
+  def testEmptyMapsVsLists: Unit = {
+    printJSONFull("{}", Map()) 
+    printJSONFull("[]", List())
+  }
   
   // Lists should be returned in the same order as specified
-  printJSON("[4,1,3,2,6,5,8,7]", JSONArray(List[Double](4,1,3,2,6,5,8,7)))
+  @Test
+  def testListOrder: Unit = 
+    printJSON("[4,1,3,2,6,5,8,7]", JSONArray(List[Double](4,1,3,2,6,5,8,7)))
 
   // Additional tests
+  @Test
+  def testAdditional: Unit =
   printJSON("{\"age\": 0}")
 
   // The library should do a proper toString representation using default and custom renderers (ticket #3605)
-  stringDiff("{\"name\" : \"va1ue\"}", JSONObject(Map("name" -> "va1ue")).toString)
-  stringDiff("{\"name\" : {\"name1\" : \"va1ue1\", \"name2\" : \"va1ue2\"}}",
-             JSONObject(Map("name" -> JSONObject(TreeMap("name1" -> "va1ue1", "name2" -> "va1ue2")))).toString)
+  @Test
+  def testDefaultAndCustomRenderers: Unit = {
+    assertEquals("{\"name\" : \"va1ue\"}", JSONObject(Map("name" -> "va1ue")).toString())
+    assertEquals("{\"name\" : {\"name1\" : \"va1ue1\", \"name2\" : \"va1ue2\"}}",
+               JSONObject(Map("name" -> JSONObject(TreeMap("name1" -> "va1ue1", "name2" -> "va1ue2")))).toString())
 
-  stringDiff("[4.0, 1.0, 3.0, 2.0, 6.0, 5.0, 8.0, 7.0]", JSONArray(List[Double](4,1,3,2,6,5,8,7)).toString)
+    assertEquals("[4.0, 1.0, 3.0, 2.0, 6.0, 5.0, 8.0, 7.0]", JSONArray(List[Double](4,1,3,2,6,5,8,7)).toString())
+  }
 
   // A test method that escapes all characters in strings
   def escapeEverything (in : Any) : String = in match {
     case s : String => "\"" + s.map(c => "\\u%04x".format(c : Int)).mkString + "\""
-    case jo : JSONObject => jo.toString(escapeEverything)
-    case ja : JSONArray => ja.toString(escapeEverything)
+    case jo : JSONObject => jo.toString(escapeEverything _)
+    case ja : JSONArray => ja.toString(escapeEverything _)
     case other => other.toString
   }
 
-  stringDiff("{\"\\u006e\\u0061\\u006d\\u0065\" : \"\\u0076\\u0061\\u006c\"}", JSONObject(Map("name" -> "val")).toString(escapeEverything))
+  @Test
+  def testEscapeEverything: Unit = 
+    assertEquals("{\"\\u006e\\u0061\\u006d\\u0065\" : \"\\u0076\\u0061\\u006c\"}", JSONObject(Map("name" -> "val")).toString(escapeEverything _))
 
-  println
 
   // from http://en.wikipedia.org/wiki/JSON
   val sample1 = """
@@ -156,9 +169,9 @@ object Test extends App {
     )
   )
 
-  
-  printJSONFull(sample1, sample1Obj)
-  println
+  @Test
+  def testSample1: Unit = 
+    printJSONFull(sample1, sample1Obj)
 
   // from http://www.developer.com/lang/jscript/article.php/3596836
   val sample2 = """
@@ -186,8 +199,9 @@ object Test extends App {
    ]
 }"""
 
-  printJSON(sample2)
-  println
+  @Test
+  def testSampl2: Unit = 
+    printJSON(sample2)
 
   // from http://json.org/example.html
   val sample3 = """
@@ -282,6 +296,7 @@ object Test extends App {
   }
 }"""
 
-  printJSON(sample3)
-  println
+  @Test
+  def testSample3 =
+    printJSON(sample3)
 }
