--- test/files/run/t0700.check
@@ -1,2 +1,0 @@
-[3.2] parsed: List(2, 2, 2)
-[3.2] parsed: List(2, 2, 2)
