--- test/files/run/t1100.check
@@ -1,4 +1,0 @@
-[1.4] error: errors are propagated
-
-aaab
-   ^
