--- test/files/run/t4929.check
@@ -1,1 +1,0 @@
-success
